% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonCI.R
\name{bonCI}
\alias{bonCI}
\title{Simultaneous Bonferroni Confidence Intervals for Lambdas}
\usage{
bonCI(df, m, alpha = 0.05)
}
\arguments{
\item{df}{, df a data frame}

\item{m}{, m the number of simultaneous intervals}

\item{alpha, }{alpha is the condifendce level for the intervals}
}
\value{

}
\description{
This function will perform simulataneous confidence interval estimates. The
estimates are corrected using the bonferroni method of correction.
}
\examples{
data <- iris[,1:3]
bonCI(data, 2)
}
