% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canCorrPlot.R
\name{canCorrPlot}
\alias{canCorrPlot}
\title{Canonical Correlation Analysis}
\usage{
canCorrPlot(df1, df2, plotType = "obs", pair = 1, labels)
}
\arguments{
\item{df1}{, df1 is a data frame with numerical columns}

\item{df2}{, df2 is a data frame with numerical columns}

\item{plotType, }{either "obs" for observation plot or "bi" for plots of the differnt variables from both data sets or "tri" for all scores}

\item{pair, }{pair tells which pair of linear combinations will make up the axes}

\item{labels, }{label all names of variables in df1 and df2 in order}
}
\value{

}
\description{
This function will perform a canonical correlation analysis between the 2 data sets entered as
arguments and plot one or all of the observations and the variables in the component space
for any of the different pairs of linear combinations.
}
\examples{
data <- rotateS21::covidWrangled
demographics <- data[,9:15]
severity <- data[,16:19]
lab <- c("Male", "Asian", "Black", "MultiRace", "NativePI", "White", "Ethnicity", "Symptomatic", "Hospital", "ICU", "Death")
canCorrPlot(demographics,severity, labels=lab)
}
