% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiPlot.R
\name{chiPlot}
\alias{chiPlot}
\title{Chi QQ Plot}
\usage{
chiPlot(df)
}
\arguments{
\item{df}{, a data frame with continuous numeric columns}
}
\value{
, a qq chi plot
}
\description{
This plot will show the ordered squared distances for each of the observations on
the y axis. On the x axis we have the chi square quantiles of (j-0.5)/n for each j.
The quantiles are calculated with the chi square distribution on p degrees of freedom
where p is the number of columns in the data frame.
}
\examples{
df <- iris
chiPlot(iris[,c(1,2,3)])

}
