% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confEllipse.R
\name{confEllipse}
\alias{confEllipse}
\title{Bivariate Quality Control Ellipse}
\usage{
confEllipse(data, vec = NULL, alpha = 0.05, formatted = TRUE)
}
\arguments{
\item{data}{, a 2 column data frame}

\item{vec}{, a test point in the form of a 2x1 bector}

\item{alpha}{, a confidence level}

\item{formatted}{, a logical argument to determine whether the data is formatted correctly}
}
\value{

}
\description{
This function has 2 main features. The first feature is the quality control ellipse
chart itself. The chart can be used to visually asses bivariate normality. The
shape and direction of the ellipse gives information about the covariance between
the 2 data columns. The second feature is test points. A
vector argument can be supplied to be plotted on the control chart. The point is
plotted in red for easy viewing. The user can easily determine whether the point
being tested falls inside the quality control ellipse.
}
\examples{

data(iris)
iris2<-iris[,c(1,2)]
confEllipse(iris2, c(6, 2.9))

}
