% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidWrangled}
\alias{covidWrangled}
\title{Covid Patient Data}
\format{
A data frame with 408612 rows and 19 variables
}
\usage{
covidWrangled
}
\description{
A dataset describing covid patients by demographic and outcome of case. All categorical variables dummy coded.
}
\keyword{datasets}
