% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenPCA.R
\name{eigenPCA}
\alias{eigenPCA}
\title{Perform an Eigen Analysis to Construct Principal Components}
\usage{
eigenPCA(df)
}
\arguments{
\item{df}{, df is a data frame of quantatative variables}
}
\value{

}
\description{
This function will allow you to construct a set of principal components using
the eigen analysis of the covariance matrix built from the input data frame. The
function will return, as a list, the covariance matrix, the linear combinations
of the principal components, the total sample variance of each component, and a
matrix containing the correlation coefficients between each linear combination and
each variable.
}
\examples{
data <- iris[,1:3]
eigenPCA(data)
}
