% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipseScatter.R
\name{ellipseScatter}
\alias{ellipseScatter}
\title{EllipseScatter}
\usage{
ellipseScatter(df, chi = 0.05)
}
\arguments{
\item{df}{, a data frame of 2 continuous variables}

\item{chi}{, the desired chi square quantile}
}
\value{
a scatter plot with an overlaid ellipse
}
\description{
This function will produce a scatterplot of the 2 variables. An ellipse determined by
the covariance matrix is overlaid on top of the scatter plot. To form the ellipse
we turn to the covariance matrix. The length of the major axis is the square root
of the first eigenvalue of the covariance matrix and the direction is determined by
the associated eigenvector. Similarly, the length of the mainor axis is the square
root of the second eigenvalue of the covariance matrix and the direction is
determined by the associated eigenvector.
}
\examples{
df <- iris
ellipseScatter(iris[,c(1,2)])

}
