% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVnorm.R
\name{MVnorm}
\alias{MVnorm}
\title{Interactive Multivariate Normality Check}
\usage{
MVnorm()
}
\value{

}
\description{
This function launches an app that allows the user to interact with the data directly
while performing normality checks.
}
\examples{
\dontrun{MVnorm()}

}
