% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcabootAG.R
\name{pcabootAG}
\alias{pcabootAG}
\title{Bootstrap Analysis of PCA}
\usage{
pcabootAG(df, alpha = 0.05, iter, covar = TRUE)
}
\arguments{
\item{df}{, df is a data frame of quantitative variables}

\item{alpha, }{alpha is a confidence level}

\item{iter, }{iter is the number of desired bootstrap iterations}

\item{covar, }{covar is a logical arg where TRUE uses covariance and FALSE uses correlation}
}
\value{
A list a confidence intervals
}
\description{
This function allows the user to make estimates for the eigenvalues and eigenvectors
of a covariance or correlation matrix. The function provides bootstrap and
Anderson-Girshick estimates for the lambdas (eigenvalues). The function provides
bootstrap estimates for the eigenvectors. The iter argument controls the number of
times the bootstrap is run, and the alpha determines the confidence level printed
to the console. Histograms for the lambdas and violin plots for the eigenvectors are
printed as well. The three types of confidence intervals are stored as lists to be
called if the results of the function are stored as an object.
}
\examples{
df <- iris[c(2,3)]
pcabootAG(df, 0.01, iter=1000, covar=TRUE )

}
