% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{Visual Profile Analysis}
\usage{
profiles(df, x, y1, y2, y3)
}
\arguments{
\item{df}{, df, a data frame with at least 4 categorical columns}

\item{x}{, x, name of a column in the data frame to use as the predictor variable}

\item{y1}{, y1, name of a binary column in the data frame to use as the first response variable}

\item{y2}{, y2, name of a binary column in the data frame to use as the second response variable}

\item{y3}{, y3, name of a binary column in the data frame to use as the third response variable}
}
\value{

}
\description{
This function will create a graph that compares the levels of a single categorical predictor variable against 3 different
categorical response variables. The x axis will be the categorical variable. The y axis will the proportion of each level of the
predictor variable that displays the response variable.
}
\examples{
data <- rotateS21::covidWrangled
profiles(data, "race", "hosp_yn_Yes", "icu_yn_Yes", "death_yn_Yes")
}
