% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProject3.R
\name{shinyProject3}
\alias{shinyProject3}
\title{Interactive Multivariate Analysis}
\usage{
shinyProject3()
}
\value{

}
\description{
This function will launch the shiny app that allows you to interact with the multivariate analysis data I've chosen for project 3.
The app goes over trends in covid severity as it relates to demographic information about each patient.
}
\examples{
\dontrun{ shinyProject3()}
}
