% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRotate.R
\name{shinyRotate}
\alias{shinyRotate}
\title{Interactive Data Cloud Rotation}
\usage{
shinyRotate()
}
\value{

}
\description{
This function launches a shiny app to interact with a data cloud. Explore the effects
of different rotations of the axes.You will be able to use the built in sample data
set or one of your own. Choose which variables to plot and see how an axis rotation
effects the covariance of the variables. Widgets are provided to allow you to change
the visual asthetics of the data presentation.
}
\examples{
\dontrun{ shinyRotate()}
}
