% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsq.R
\name{tsq}
\alias{tsq}
\title{Tsq Chart}
\usage{
tsq(df, chi = 0.05)
}
\arguments{
\item{df}{, a data frame}

\item{chi}{, the desired alpha level}
}
\value{
, a tsq chart
}
\description{
This chart will give an idea of which observations are out of control. The chi square
distance of each observation is plotted on the y axis, and the ordering of the x axis
comes from time series ordering. The upper control limit (percent defined by the chi
argument) is also charted. Observations above this line are considered out of control
and should be examined further.
}
\examples{


}
