% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniControl.R
\name{uniControl}
\alias{uniControl}
\title{Univariate Control Charts}
\usage{
uniControl(df)
}
\arguments{
\item{df}{, a data frame with continuous numerical columns}
}
\value{

}
\description{
This function will produce univariate control charts to visually assess which
observations are out of control with regards to a single variable. The variable
being assessed for control is plotted on the y-axis for each observation. The ordering
of the x axis comes from time series ordering. The upper and lower control limits
(percent defined by the alpha argument) is also charted. Points that fall above or
below the control limit lines are considered out of control
}
\examples{
df <- iris
uniControl(df[,c(2,3)])

}
