% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confEllipse.R
\name{confEllipse}
\alias{confEllipse}
\title{Bivariate Confidence Ellipse for Multivariate Normal Data}
\usage{
confEllipse(data, vec, alpha = 0.05, formatted = TRUE)
}
\arguments{
\item{data}{, a data frame}

\item{vec}{, a test point}

\item{alpha}{, a confidence level}

\item{formatted}{, a logical argument to determine whether the data is formatted correctly}
}
\value{

}
\description{
Bivariate Confidence Ellipse for Multivariate Normal Data
}
\examples{
data(iris)
iris2<-iris[,c(1,2)]
confEllipse(iris2, c(6, 2.9))

}
