% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{Visual Profile Analysis}
\usage{
profiles(df, x1, y1, y2, y3)
}
\arguments{
\item{df}{, df, a data frame with at least 4 categorical columns}

\item{x1}{, x, index of a binary (0,1) column in the data frame to use as the predictor variable}

\item{y1}{, y1, index of a binary (0,1) column in the data frame to use as the first response variable}

\item{y2}{, y2, index of a binary (0,1) column in the data frame to use as the second response variable}

\item{y3}{, y3, index of a binary (0,1) column in the data frame to use as the third response variable}
}
\value{

}
\description{
This function will create a graph that compares the levels of a single categorical predictor variable against 3 different
categorical response variables. The x axis will be the categorical variable. The y axis will the proportion of each level of the
predictor variable that displays the response variable.
}
\examples{
dataframe <- rotateS21::covidWrangled
profiles(dataframe, 3, 17, 18, 19)
}
